+++ src/Domain/Model/PurchaseOrder/QuantityReceived.php
@@ -8,11 +8,11 @@ use Domain\Model\ReceiptNote\ReceiptQuantity;
 final class QuantityReceived
 {
     /**
-     * @var float
+     * @var int
      */
     private $quantity;
 
-    public function __construct(float $quantity)
+    public function __construct(int $quantity)
     {
         if ($quantity < 0) {
             throw new \InvalidArgumentException('Quantity received should be larger than or equal to 0.');
@@ -21,18 +21,18 @@ final class QuantityReceived
         $this->quantity = $quantity;
     }
 
-    public function asFloat(): float
+    public function asInt(): int
     {
         return $this->quantity;
     }
 
     public function add(ReceiptQuantity $quantity): QuantityReceived
     {
-        return new self($this->quantity + $quantity->asFloat());
+        return new self($this->quantity + $quantity->asInt());
     }
 
     public function subtract(ReceiptQuantity $quantity): QuantityReceived
     {
-        return new self($this->quantity - $quantity->asFloat());
+        return new self($this->quantity - $quantity->asInt());
     }
 }
